feat: split Blood & Body into separate Age, Blood, Body, and Eyes pages
Each health domain now has its own dedicated tab and sidebar entry
with distinct icons, improving navigation and reducing page bloat.
