+++ .changelog/v1.17.x.md
@@ -1,0 +1,13 @@
+# Release v1.17.x - Changelog
+
+Released: YYYY-MM-DD
+
+## Overview
+
+Split monolithic "Blood & Body" tab into separate focused pages for better navigation.
+
+## Features
+
+- **Separate MeatSpace pages** — split the "Blood & Body" tab into 4 independent pages: Age (epigenetic), Blood (tests), Body (composition), and Eyes (prescriptions)
+- **Dedicated sidebar nav entries** — each health domain now has its own sidebar link with distinct icons (Clock, HeartPulse, Scale, Eye)
+- **Correct overview tile navigation** — overview health tiles now link directly to their respective pages instead of all pointing to Blood
